pyproject.toml
[project]
name = "divyadrishti"
version = "0.1.0"
description = "Multi-source public intelligence platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery[redis]>=5.3",
    "duckdb>=1.0",
    "fastapi>=0.115",
    "httpx>=0.27",
    "numpy>=1.24",
    "openai>=1.0",
    "pydantic>=2.9",
    "python-dotenv>=1.0",
    "pydantic-settings>=2.0",
    "redis>=5.0",
    "sse-starlette>=1.0",
    "uvicorn>=0.30",
]

[dependency-groups]
dev = [
    "fakeredis>=2.0",
    "mypy>=1.11",
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["libs", "agents", "apps"]

[tool.ruff]
target-version = "py311"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["slow: marks tests that load ML models (deselect with '-m \"not slow\"')"]
